cmd/observe: move TestAgentEventSubTypeMap to other agent events tests
